ItemCard and a hardcoded Item List added
Hardcoded a list with 4 items that represent one collection on focus on main page.
It is just a demo and can be worked out later.
Some basic css added in separate files in case we decide to row back or not include it, also to ensure easier teamwork with less conflicts.