Banner created and CSS files moved to main.scss
A banner with 3 images is created and available on Homepage only.

Known problem is the scaling on the image on full screen on a desktop. The image gets cut accordingly and it does not look good on big screen.

Same problem occurs on small iamges when they become 200px instead of 450 for the items. Then the image is cut. A potential solution could be  different sizes of images, that should be used on different medias.

However, it is out of scope for now and it might be solved when its priority become bigger, or the team has time over.