Get products by user only backend
The database is changed a bit and a .sql is provided (obs - No data in - just structure!). The SQL works with MySQL.

The backend does not take into consideration which user is currently logged in to provide them with an info only about their products which mean that if an inlogged user checks another users id they will see their products as well.

However, this is only the backend and it is not connected to the fronend mostly because we have no logic to take the inlogged users ID and there is a lot about protecting the data.

The obnoxious logos are added in even if we are not going to use them. My code uses them and you can see the result in trello.