The backend of the filter by product quantity less than 0
A new endpints /filter that can take maxQuantity value. I was thinking of scalability. I have the frontend but I do not want to upload it for now

By filtering by maxQuality we allow the admin to change their mind - they dont only want to see the 0 and 1 quantity items but maybe they want to check every product that is under 20... or 0 and less.