Simple checkbox added
One solution with a simple checkbox is presented together with the more complex filter.

On checked it will show only the filter with a default value of Less than 2. Unchecked shows everything

A small change is made to the input field to take the input number as a desired max quantity and not Less than that quantity.