Filter is implemented
Filter the displayed results in admin view.

The admin can write what the max quantity can be, providing scalability. By clicking filter with nothing in the input field the filtered quantities will give results for Less than 2.

The checkbox handler is still in the code in case we decide to use it for something else.