Filter - show products by inlogged admin
New functionality - filter that shows the items added by the currently logged in admin.

The filter does not work in combination with the filter for Max quantity - if one of them is chosen the other one is not active.

For the change I had to get the use ID, thus I communicate with the context.