Default picture in case of no item images and a toast adjust
If a picture doesn'т have an image it will display one from the internet by url. That ensures design consistency in cards and inside image - so the broken picture will not be Previewed.

The position of the toste messages is changed to bottom left and the duration is reduced to 2 sec instead of 3.